# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "tinrist123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/post/tags/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get tags have foreign key with post ",
                "description": "Returns Tags By id Post",
                "operationId": "getProjectById33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/total": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get total of post",
                "description": "Return total of post",
                "operationId": "getTotalPost",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/v1/post/category/image/{idCate}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get post have id category which is foreign key with Post ",
                "description": "Returns post By id Category foreign key Post",
                "operationId": "getPost-Related-Category",
                "parameters": [
                    {
                        "name": "idCate",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/related/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get related post have id category which is foreign key with Post ",
                "description": "Returns related post By id Category foreign key Post",
                "operationId": "getPost-Related-Post-By-Post-Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/new/page/{page}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get new post per page",
                "description": "Returns post per page",
                "operationId": "getnewPost-show-per-page",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get Post By Id",
                "description": "Returns Post",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/user/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get user by id Post ",
                "description": "Returns User",
                "operationId": "getUserByIdPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/comment/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get comment by id Post ",
                "description": "Returns comment",
                "operationId": "getCommentgoryByIdPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/category/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get category by id Post ",
                "description": "Returns category",
                "operationId": "getCategoryByIdPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/post/tag/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get tag by id Post ",
                "description": "Returns tag",
                "operationId": "getTagByIdPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}